Don’t bubble events.
This is confusing when you have multiple launchers nested (with the upcoming legacy launcher), and it’s not very helpful anyway because most of the events are very specific to a particular iframe launcher.

Only bubble track and error, because for them it’s useful to have a global event listener.